pyproject.toml
[tool.poetry]
name        = "realwin_api"
version     = "0.1.0"
description = ""
readme      = "README.md"
authors     = ["Zohaib"]
package-mode = false

[tool.poetry.dependencies]
python              = ">=3.8,<3.12"
alembic             = "1.12.1"
asyncpg             = "0.29.0"
email-validator     = "2.1.0.post1"
fastapi             = "^0.116.1"
passlib             = "1.7.4"
psycopg2-binary     = "^2.9.9"
pydantic-settings   = "2.1.0"
python-jose         = "3.3.0"
sqlalchemy          = "^2.0.0"
uvicorn             = "^0.35.0"
greenlet            = "^3.1.1"
bcrypt              = "^4.0.1"
python-dotenv       = "^1.1.1"
httpx               = "^0.28.1"
fastapi-mail        = "1.4.0"
boto3 = "^1.38.23"
websockets = "^15.0.1"
aiohttp = "^3.12.12"
openai = "^1.102.0"
pytz = "^2025.2"
bs4 = "^0.0.2"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
werkzeug = "^3.1.3"
jinja2 = "^3.1.6"
apscheduler = "^3.11.0"
# M.A.X. Agent Dependencies
langgraph = "^0.6.7"
langchain = "^0.3.27"
langchain-openai = "^0.3.32"
langchain-pinecone = "^0.2.12"
tabulate = "^0.9.0"
stripe = "11.2.0"
# Enhanced Cricket Intelligence Dependencies
feedparser = "^6.0.10"
selenium = "^4.15.2"
lxml = "^4.9.3"
requests-html = "^0.10.0"
colorama = "^0.4.6"
# Real-time web search capabilities
tavily-python = "^0.3.0"
aiofiles = "^25.1.0"
# MAX Ultimate System Dependencies
scikit-learn = "^1.3.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
asyncio-mqtt = "^0.16.1"
websocket-client = "^1.6.0"
aioredis = "^2.0.0"
python-dateutil = "^2.8.2"
schedule = "^1.2.0"
emoji = "^2.8.0"
nltk = "^3.8.0"
textblob = "^0.17.1"
regex = "^2023.8.8"
tqdm = "^4.66.0"
retrying = "^1.3.4"
cachetools = "^5.3.0"
# Additional ML and AI dependencies
torch = "^2.0.0"
transformers = "^4.30.0"
# Enhanced web scraping
fake-useragent = "^1.4.0"
cloudscraper = "^1.2.71"
# Data processing
xlsxwriter = "^3.1.0"
openpyxl = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pre-commit     = "3.5.0"
pytest         = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires    = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
